Move everything into `definiti.core` instead of `definiti`.
