+++ src/main/scala/definiti/core/utils/ParserUtils.scala
@@ -1,6 +1,6 @@
-package definiti.utils
+package definiti.core.utils
 
-import definiti.{ParameterDefinition, Position, Range, TypeReference, Variable}
+import definiti.core.{ParameterDefinition, Position, Range, TypeReference, Variable}
 import org.antlr.v4.runtime.ParserRuleContext
 import org.antlr.v4.runtime.tree.TerminalNode
