+++ src/test/scala/definiti/core/api/ASTHelperSpec.scala
@@ -1,7 +1,7 @@
-package definiti.api
+package definiti.core.api
 
-import definiti.parser.{ASTValidation, Valid}
-import definiti.{And, AttributeCall, AttributeDefinition, BooleanValue, CombinedExpression, Condition, Equal, MethodCall, NativeClassDefinition, NativeMethodDefinition, NumberValue, Position, QuotedStringValue, Range, TypeReference, Variable}
+import definiti.core.parser.{ASTValidation, Valid}
+import definiti.core.{And, AttributeCall, AttributeDefinition, BooleanValue, CombinedExpression, Condition, Equal, MethodCall, NativeClassDefinition, NativeMethodDefinition, NumberValue, Position, QuotedStringValue, Range, TypeReference, Variable}
 import org.scalatest.{FlatSpec, Matchers}
 
 class ASTHelperSpec extends FlatSpec with Matchers {
